[Review] Localizable 추가 및 ErrorType 코드 분리
